%%
\name{Conversion-methods}
%%
%%
%%
\docType{methods}
%%
%%
%%
\alias{coerce}
\alias{coerce-methods}
%%
\alias{coerce,IPv4,character-method}
\alias{coerce,IPv6,character-method}
%%
%%
%%
\alias{as.character}
\alias{as.character-methods}
%%
\alias{as.character,IPv4-method}
\alias{as.character,IPv4r-method}
\alias{as.character,IPv6-method}
\alias{as.character,IPv6r-method}
\alias{as.character,IP-method}
\alias{as.character,IPr-method}
%%
\alias{as.character,host-method}
%%
%%
%%
\alias{format}
\alias{format.IPv4}
\alias{format.IPv6}
\alias{format.IP}
\alias{format.IPv4r}
\alias{format.IPv6r}
\alias{format.IPr}
%%
%%
%%
\alias{toString}
\alias{toString.IPv4}
\alias{toString.IPv6}
\alias{toString.IP}
\alias{toString.IPv4r}
\alias{toString.IPv6r}
\alias{toString.IPr}
%%
%%
%%
%%
\alias{as.numeric}
%%
\alias{as.numeric-methods}
\alias{as.numeric,IPv4-method}
\alias{as.numeric,IPv4r-method}
\alias{as.numeric,IPv6-method}
\alias{as.numeric,IPv6r-method}
\alias{as.numeric,IP-method}
%%
%%
%%
% \alias{coerce}
% \alias{coerce-method}
% \alias{coerce,IPv4,character-method}
% \alias{coerce,IPv6,character-method}
%%
%%
%%
\title{ Methods for converting IP objects to other representations }
%%
\description{
  Methods for converting IP objects to other representations
  %%
  % \itemize{
    % %%
    % \item \code{as.character}
  % }
}
% \usage{
  % as.character(x,\ldots)
% }
%%
\arguments{
  \item{x}{ an atomic vector of class 'IPv4', 'IPv6' or 'IP'}
  \item{\ldots}{not used}
}
%%
% \details{}
%%
%%
%%
% \examples{
  
% }
