% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getipintel-contact-info.R
\name{getipintel_contact_info}
\alias{getipintel_contact_info}
\title{Get or set GETIPINTEL_CONTACT_INFO value}
\usage{
getipintel_contact_info(force = FALSE)
}
\arguments{
\item{force}{Force setting a new GetIPIntel contact info string for the current environment?}
}
\value{
atomic character vector containing the PGetIPIntel contact info string
}
\description{
The API wrapper functions in this package all rely on a GetIPInfo API
contact info string residing in the environment variable \code{GETIPINTEL_CONTACT_INFO}.
The easiest way to accomplish this is to set it
in the \code{.Renviron} file in your home directory.
}
\references{
\url{https://getipintel.net/}
}
\author{
Bob Rudis (bob@rud.is)
}
