% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iphub-api-key.R
\name{iphub_api_key}
\alias{iphub_api_key}
\title{Get or set IPHUB_API_KEY value}
\usage{
iphub_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new IPHub key for the current environment?}
}
\value{
atomic character vector containing the IPHub api key
}
\description{
The API wrapper functions in this package all rely on a PacketTotal API
key residing in the environment variable \code{IPHUB_API_KEY}.
The easiest way to accomplish this is to set it
in the \code{.Renviron} file in your home directory.
}
\references{
\url{https://iphub.info/api}
}
\author{
Bob Rudis (bob@rud.is)
}
