% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxycheck-api-key.R
\name{proxycheck_api_key}
\alias{proxycheck_api_key}
\title{Get or set PROXYCHECK_API_KEY value}
\usage{
proxycheck_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new ProxyCheck key for the current environment?}
}
\value{
atomic character vector containing the ProxyCheck api key
}
\description{
The API wrapper functions in this package all rely on a PacketTotal API
key residing in the environment variable \code{PROXYCHECK_API_KEY}.
The easiest way to accomplish this is to set it
in the \code{.Renviron} file in your home directory.
}
\references{
\url{https://proxycheck.io/api/#introduction}
}
\author{
Bob Rudis (bob@rud.is)
}
