\name{RIP.classifiers}
\alias{RIP.classifiers}
\docType{data}
\title{
Provided RIP master classifier
}
\description{
A list containing the RIP master classifier used to predict regulatory interactions (RIs) by ten features. See source for more details on the features and the classifier.
}
\usage{data(RIP.classifiers)}
\source{
Bauer et al. (2011) Bioinformatics 27:2239-47 - RIP: the regulatory interaction predictor--a machine learning-based approach for predicting target genes of transcription factors.
}
\examples{
data(RIP.classifiers)
}
\keyword{datasets}
