% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rIP-package.R
\docType{package}
\name{rIP}
\alias{rIP}
\alias{rIP-package}
\title{Detects Fraud in Online Surveys by Tracing, Scoring, and Visualizing IP Addresses}
\description{
Takes an array of IPs and the keys for the services the user wishes to use (IP Hub, IP Intel, and Proxycheck), 
    and passes these to all respective APIs. Returns a dataframe with the IP addresses (used for merging), 
    country, ISP, labels for non-US IP Addresses, VPS use, and recommendations for blocking. The package 
    also provides optional visualization tools for checking the distributions. Additional
    functions are provided to call each discrete API endpoint.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://joss.theoj.org/papers/10.21105/joss.01285}
  \item Report bugs at \url{https://github.com/MAHDLab/rIP/issues}
}

}
\author{
\strong{Maintainer}: Ryan Kennedy \email{rkennedy@uh.edu}

Authors:
\itemize{
  \item Philip Waggoner \email{philip.waggoner@gmail.com} (0000-0002-7825-7573)
}

Other contributors:
\itemize{
  \item Scott Clifford [contributor]
  \item Bob Rudis \email{bob@rud.is} (0000-0001-5670-2640) [contributor]
}

}
\keyword{internal}
