\name{RIP.RIpredictions}
\alias{RIP.RIpredictions}
\docType{data}
\title{
General RIs predicted from the CAMDA2007 dataset
}
\description{
These RIs were predicted from 4064 selected gene expression microarrays of the CAMDA2007 dataset (ArrayExpress accession E-TABM-185). They are general in the sense that the data comprised 76 different experiments (i.e. a broad spectrum of different primary tissue and tumor samples) on which the correlation meta-analysis had been conducted.
}
\source{
Bauer et al. (2011) Bioinformatics 27:2239-47 - RIP: the regulatory interaction predictor--a machine learning-based approach for predicting target genes of transcription factors.

\url{http://www.ichip.de/software/RIP.html}
}
\references{
Affymetrix hgu133a annotation used: \url{http://www.bioconductor.org/packages/2.4/AnnotationData.html}
}
\examples{
str(RIP.RIpredictions)
}
\keyword{predicted RIs}
